Moving support for reading parameter names from javadoc and for generating artificial parameter names to the ParameterNameProviderImpl.
